src/utils/config.py
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

API_KEY = os.getenv("GEMINI_API_KEY")

if not API_KEY:
    raise ValueError("Google Gemini API Key is missing. Set GEMINI_API_KEY in the .env file.")