requirements.txt
google-generativeai
python-dotenv
pyperclip
keyboard
plyer
tk